Cambios para permitir distintos handlers al logger
