+++ php/toba_autoload.php
@@ -433,7 +433,10 @@ class toba_autoload
 		'toba_interface_sesion' => 'nucleo/lib/toba_interface_sesion.php',
 		'toba_interface_usuario' => 'nucleo/lib/toba_interface_usuario.php',
 		'toba_js' => 'nucleo/lib/toba_js.php',
-		'toba_basic_logger' => 'nucleo/lib/toba_basic_logger.php',
+		'toba_basic_logger' => 'nucleo/lib/log_handlers/toba_basic_logger.php',
+		'logger' => 'nucleo/lib/log_handlers/logger.php',
+		'toba_handler_log' => 'nucleo/lib/log_handlers/toba_handler_log.php',
+		'toba_ws_handler_log' => 'nucleo/lib/log_handlers/toba_ws_handler_log.php',
 		'toba_logger' => 'nucleo/lib/toba_logger.php',
 		'toba_logger_ws' => 'nucleo/lib/toba_logger_ws.php',
 		'toba_mail' => 'nucleo/lib/toba_mail.php',
